Reorganize new types
